+++ test/client_test.dart
@@ -1,39 +1,19 @@
-import 'package:fresh_dio/fresh_dio.dart';
+import 'package:flutter/widgets.dart';
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:test/test.dart';
 import 'package:wiseclient/wiseclient.dart';
 
 void main() {
+  WidgetsFlutterBinding.ensureInitialized();
+  FlutterSecureStorage.setMockInitialValues({});
   group('A group of basic tests to test functionality', () {
     final wiseOptions = WiseOptions.baseWithLocale(
       url: 'https://jsonplaceholder.typicode.com/',
       locale: 'en',
     );
     final awesome = WiseClient(
-      wiseInterceptors: WiseInterceptor.values,
+      wiseInterceptors: [WiseInterceptor.logging, WiseInterceptor.error],
       options: wiseOptions,
-      refreshFunction: (
-        token,
-        client,
-      ) async {
-        try {
-          final newToken = await client.post<OAuth2Token>(
-            '/oauth/token',
-            data: {
-              'client_id': 'clientId',
-              'client_secret': 'clientSecret',
-              'grant_type': 'refresh_token',
-              'refresh_token': token?.refreshToken,
-            },
-          );
-
-          return OAuth2Token(
-            accessToken: newToken.data?.accessToken ?? '',
-            refreshToken: newToken.data?.refreshToken,
-          );
-        } catch (e) {
-          throw RevokeTokenException();
-        }
-      },
     );
 
     test('Client is native', () {
@@ -69,6 +49,33 @@ void main() {
       const path = 'todos/1';
       final something = await awesome.wGet(path);
       expect(something, isMap);
+    });
+
+    test('Can add headers to requests', () async {
+      const path = 'todos/1';
+      const additionalHeaders = {'testHeader': 'testValue'};
+      final something = await awesome.wGet(
+        path,
+        extraHeaders: additionalHeaders,
+      );
+      expect(something, isMap);
+    });
+
+    test('Does not remember headers', () async {
+      const path = 'todos/1';
+      const additionalHeaders1 = {'testHeader1': 'testValue1'};
+      final firstResult = await awesome.wGet(
+        path,
+        extraHeaders: additionalHeaders1,
+      );
+
+      const additionalHeaders2 = {'testHeader2': 'testValue2'};
+      final secondResult = await awesome.wGet(
+        path,
+        extraHeaders: additionalHeaders2,
+      );
+      expect(firstResult, isMap);
+      expect(secondResult, isMap);
     });
 
     test('Wise post works', () async {